requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-background-tasks-updated==1.2.7
django-compat==1.0.15
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
